(#252) Fix hot reloading the sounds
