+++ src/something_game.cpp
@@ -271,7 +271,7 @@ void Game::update(float dt)
                 projectile->kill();
                 entity->lives -= ENTITY_PROJECTILE_DAMAGE;
 
-                mixer.play_sample(damage_enemy_sample);
+                mixer.play_sample(assets.sounds[assets.get_sound_by_id_or_panic("OOF_SOUND"_sv).unwrap].unwrap);
                 if (entity->lives <= 0) {
                     for (size_t i = 0; i < entity->dirt_blocks_count; ++i) {
                         const float ITEMS_DROP_PROXIMITY = 50.0f;
@@ -291,7 +291,7 @@ void Game::update(float dt)
                     }
 
                     entity->kill();
-                    mixer.play_sample(kill_enemy_sample);
+                    mixer.play_sample(assets.sounds[assets.get_sound_by_id_or_panic("CRUNCH_SOUND"_sv).unwrap].unwrap);
                 } else {
                     entity->vel += normalize(projectile->vel) * ENTITY_PROJECTILE_KNOCKBACK;
                     entity->flash(ENTITY_DAMAGE_FLASH_COLOR);
@@ -320,19 +320,19 @@ void Game::update(float dt)
                         case ITEM_HEALTH: {
                             entity->lives = min(entity->lives + ITEM_HEALTH_POINTS, ENTITY_MAX_LIVES);
                             entity->flash(ENTITY_HEAL_FLASH_COLOR);
-                            mixer.play_sample(item->sound);
+                            mixer.play_sample(assets.sounds[item->sound.unwrap].unwrap);
                             item->type = ITEM_NONE;
                         } break;
 
                         case ITEM_DIRT_BLOCK: {
                             entity->dirt_blocks_count += 1;
-                            mixer.play_sample(item->sound);
+                            mixer.play_sample(assets.sounds[item->sound.unwrap].unwrap);
                             item->type = ITEM_NONE;
                         } break;
 
                         case ITEM_ICE_BLOCK: {
                             entity->ice_blocks_count += 1;
-                            mixer.play_sample(item->sound);
+                            mixer.play_sample(assets.sounds[item->sound.unwrap].unwrap);
                             item->type = ITEM_NONE;
                         } break;
                         }
@@ -465,7 +465,7 @@ void Game::entity_shoot(Entity_Index entity_index)
                     entity_index);
                 entity->cooldown_weapon = ENTITY_COOLDOWN_WEAPON;
 
-                mixer.play_sample(entity->shoot_sample);
+                mixer.play_sample(assets.sounds[entity->shoot_sample.unwrap].unwrap);
             }
         } break;
